% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_PCR_Test.R
\docType{data}
\name{A_PCR_Test}
\alias{A_PCR_Test}
\title{A_PCR_Test}
\format{
An object of class \code{data.frame} with 66 rows and 5 columns.
}
\usage{
A_PCR_Test
}
\description{
A_PCR_Test is the test set of the A_PCR model. It consists of Age, MeanDNE.Apex, Proportion.DNEunder0.0001, Proportion.DNEover0.6, IQRDNE.Whole. The number of rows is 66.
}
\keyword{datasets}
