% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_PCR_Train.R
\docType{data}
\name{A_PCR_Train}
\alias{A_PCR_Train}
\title{A_PCR_Train}
\format{
An object of class \code{data.frame} with 269 rows and 5 columns.
}
\usage{
A_PCR_Train
}
\description{
A_PCR_Train is the train set of the A_PCR model. It consists of Age, MeanDNE.Apex, Proportion.DNEunder0.0001, Proportion.DNEover0.6, IQRDNE.Whole. The number of rows is 269.
}
\keyword{datasets}
