% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apex.R
\docType{data}
\name{Apex}
\alias{Apex}
\title{Surface mesh of apex of auricular surface.}
\format{
An object of class \code{mesh3d} of length 4.
}
\usage{
data(Apex)
}
\description{
Surface mesh of apex of auricular surface.
}
\examples{
PCQDA_output <- PCQDA_result(WholeSurface,Apex)
PCR_output <- PCR_result(WholeSurface,Apex)
PCLR_output <- PCLR_result(WholeSurface,Apex)
}
\keyword{datasets}
