% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PCLR.R
\docType{data}
\name{data_PCLR}
\alias{data_PCLR}
\title{Database to create DNE_PCLR method}
\format{
An object of class \code{data.frame} with 682 rows and 6 columns.
}
\usage{
data_PCLR
}
\description{
data_PCLR with 6 variables: Cluster, Total DNE/Total Polygon Faces, Mean value of DNE on Apex, IQR of DNE on Whole surface, Median value of DNE on Whole surface, Mean value of DNE on Convex surface
}
\details{
The database was obtained by calculating the DNE on the auricular surface from European collections.
The specimens are belongs to  the Louis Lopes Collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan.
}
\keyword{datasets}
