% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PCQDA.R
\docType{data}
\name{data_PCQDA}
\alias{data_PCQDA}
\title{Databases to create DNE_PCQDA method.}
\format{
An object of class \code{data.frame} with 627 rows and 5 columns.
}
\usage{
data_PCQDA
}
\description{
data_PCQDA with 5 variables: Cluster, Proportion having DNE under 0.0001, Proportion having DNE over 0.6, Mean value of DNE on Apex, Total DNE/Total Polygon Faces
}
\details{
The databases was obtained by calculating the DNE on the auricular surface from European collections.
The specimens are belongs to  the Louis Lopes collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan.
}
\keyword{datasets}
