% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{data}
\name{Distribution}
\alias{Distribution}
\title{Distribution}
\format{
An object of class \code{data.frame} with 1225 rows and 3 columns.
}
\usage{
Distribution
}
\description{
It shows the sex and age distribution of the multi-populational sample.
}
\keyword{datasets}
