% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstObvervation.R
\docType{data}
\name{FirstObservation}
\alias{FirstObservation}
\title{FirstObservation}
\format{
An object of class \code{data.frame} with 140 rows and 10 columns.
}
\usage{
FirstObservation
}
\description{
FirstObservation is the raw data of DNE variables obtained from the first
observation of the Southeast Asian sample. It consists of MeanDNE.Apex, MedianDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole, IQRDNE.Whole, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001 and Proportion.DNEover0.6.
}
\keyword{datasets}
