% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSDNE-package.R
\docType{package}
\name{JSDNE-package}
\alias{JSDNE}
\alias{JSDNE-package}
\title{JSDNE: Estimating the Age using Auricular Surface by DNE}
\description{
The age is estimated by calculating the Dirichlet Normal Energy (DNE) on the whole auricular surface and the apex of the auricular surface. It involves three estimation methods: principal component discriminant analysis (PCQDA), principal component regression analysis (PCR), and principal component logistic regression analysis (PCLR) methods. The package is created with the data from the Louis Lopes Collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan.
}
\author{
\strong{Maintainer}: Jisun Jang \email{jisun.jang.19@ucl.ac.uk}

}
\keyword{internal}
