% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_PCR_Train.R
\docType{data}
\name{M_PCR_Train}
\alias{M_PCR_Train}
\title{M_PCR_Train}
\format{
An object of class \code{data.frame} with 953 rows and 7 columns.
}
\usage{
M_PCR_Train
}
\description{
M_PCR_Train is the train set of the M_PCR model. It consists of Age, MeanDNE.Apex, MedianDNE.Apex, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001, Population. The number of rows is 953.
}
\keyword{datasets}
