% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiData.R
\docType{data}
\name{MultiData}
\alias{MultiData}
\title{MultiData}
\format{
An object of class \code{data.frame} with 1225 rows and 7 columns.
}
\usage{
MultiData
}
\description{
MultiData is the combined data of M_PCR train and M_PCR test sets. It consists of Age, MeanDNE.Apex, MedianDNE.Apex, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001 and Population.
}
\keyword{datasets}
