% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCLR_result.R
\name{PCLR_result}
\alias{PCLR_result}
\title{Estimating the age using DNE_PCLR method}
\usage{
PCLR_result(x, y)
}
\arguments{
\item{x}{the name of inputted ply file of the whole auricular surface}

\item{y}{the name of inputted ply file of the apex of the auricular surface}
}
\value{
estimated result gets printed to the console
}
\description{
DNE_PCLR method is the principal component logistic regression analysis (PCLR) method using the Dirichlet Normal Energy (DNE).
This method involves 2 age groups to distinguish if the specimen is over 63 or under 67.
The function automatically calculates the DNE on the auricular surface.
It provides the estimated age group and age range of the estimated age group.
}
