% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCLR_Test.R
\docType{data}
\name{PCLR_Test}
\alias{PCLR_Test}
\title{PCLR_Test}
\format{
An object of class \code{data.frame} with 191 rows and 7 columns.
}
\usage{
PCLR_Test
}
\description{
PCR_Train is the test set of the PCR model. It consists of Age, Cluster1,
MeanDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole, IQRDNE.Whole
and MeanDNE.Convex. The number of rows is 191.
}
\keyword{datasets}
