% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCLR_Train.R
\docType{data}
\name{PCLR_Train}
\alias{PCLR_Train}
\title{PCLR_Train}
\format{
An object of class \code{data.frame} with 699 rows and 7 columns.
}
\usage{
PCLR_Train
}
\description{
PCLR_Train is the train set of the PCR model. It consists of Age, Cluster1,
MeanDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole, IQRDNE.Whole and
MeanDNE.Convex. The number of rows is 699.
}
\keyword{datasets}
