% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCQDA_result.R
\name{PCQDA_result}
\alias{PCQDA_result}
\title{Estimating the age using DNE_PCQDA method}
\usage{
PCQDA_result(x, y)
}
\arguments{
\item{x}{the name of inputted ply file of the whole auricular surface}

\item{y}{the name of inputted ply file of the apex of the auricular surface}
}
\value{
estimated result gets printed to the console
}
\description{
DNE_PCQDA method is the principal component quadratic discriminant analysis (PCQDA) method using the Dirichlet Normal Energy (DNE).
This method involves 4 age groups.
The function automatically calculates the DNE on the auricular surface.
It provides the estimated age group and age range of the estimated age group.
}
