% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCQDA_Test.R
\docType{data}
\name{PCQDA_Test}
\alias{PCQDA_Test}
\title{PCQDA_Test}
\format{
An object of class \code{data.frame} with 186 rows and 6 columns.
}
\usage{
PCQDA_Test
}
\description{
PCQDA_Test is the test set of the PCQDA model. It consists of Cluster2, Age,
MeanDNE.Apex, TotalDNE.TotalPolygonFaces, Proportion.DNEunder0.0001, and
Proportion.DNEover0.6. The number of rows is 186.
}
\keyword{datasets}
