% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCQDA_Train.R
\docType{data}
\name{PCQDA_Train}
\alias{PCQDA_Train}
\title{PCQDA_Train}
\format{
An object of class \code{data.frame} with 704 rows and 6 columns.
}
\usage{
PCQDA_Train
}
\description{
PCQDA_Train is the train set of the PCQDA model. It consists of Cluster2,
Age, MeanDNE.Apex, TotalDNE.TotalPolygonFaces, Proportion.DNEunder0.0001,
and Proportion.DNEover0.6. The number of rows is 704.
}
\keyword{datasets}
