% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR_result.R
\name{PCR_result}
\alias{PCR_result}
\title{Estimating the age using PCR method}
\usage{
PCR_result(x, y)
}
\arguments{
\item{x}{the name of inputted ply file of the whole auricular surface}

\item{y}{the name of inputted ply file of the apex of the auricular surface}
}
\value{
estimated result gets printed to the console
}
\description{
DNE_PCR method is the principal component linear regression analysis (PCR) method using the Dirichlet Normal Energy (DNE).
The function automatically calculates the DNE on the auricular surface
It provides the estimated age and doubled standard errors (2SE).
}
