% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR_Test.R
\docType{data}
\name{PCR_Test}
\alias{PCR_Test}
\title{PCR_Test}
\format{
An object of class \code{data.frame} with 188 rows and 6 columns.
}
\usage{
PCR_Test
}
\description{
PCR_Train is the test set of the PCR model. It consists of Age,
MeanDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MeanDNE.Convex
and Proportion.DNEunder0.0001. The number of rows is 188.
}
\keyword{datasets}
