% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR_Train.R
\docType{data}
\name{PCR_Train}
\alias{PCR_Train}
\title{PCR_Train}
\format{
An object of class \code{data.frame} with 702 rows and 6 columns.
}
\usage{
PCR_Train
}
\description{
PCR_Train is the train set of the PCR model. It consists of Age,
MeanDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MeanDNE.Convex
and Proportion.DNEunder0.0001. The number of rows is 702.
}
\keyword{datasets}
