% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawData.R
\docType{data}
\name{RawData}
\alias{RawData}
\title{RawData}
\format{
An object of class \code{data.frame} with 890 rows and 15 columns.
}
\usage{
RawData
}
\description{
It is the raw data of the DNE variables obtained from the European sample
to develop the PCQDA, PCLR, and PCR models. It consists of Age, Cluster1, Cluster2, Collection, Sex, MeanDNE.Apex, MedianDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole, IQRDNE.Whole, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001 and Proportion.DNEover0.6.
}
\keyword{datasets}
