% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThaiData.R
\docType{data}
\name{ThaiData}
\alias{ThaiData}
\title{ThaiData}
\format{
An object of class \code{data.frame} with 335 rows and 14 columns.
}
\usage{
ThaiData
}
\description{
ThaiData is the raw data of the DNE variables obtained from the Southeast Asian
sample to develop the A_PCR and M_PCR models. It consists of Age, MeanDNE.Apex, MedianDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole, IQRDNE.Whole, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001, Proportion.DNEover0.6, PCQDA_Estimated (results of PCQDA model) and PCR_pred (results of PCR model) , PCLR_pred (results of PCLR model).
}
\keyword{datasets}
