% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendLine.R
\docType{data}
\name{TrendLine}
\alias{TrendLine}
\title{TrendLine}
\format{
An object of class \code{data.frame} with 1225 rows and 12 columns.
}
\usage{
TrendLine
}
\description{
TrendLine is the combined data of European and Southeast Asian samples.
It consists of Age, MeanDNE.Apex, MedianDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole, IQRDNE.Whole, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001, Proportion.DNEover0.6 and Population.
}
\keyword{datasets}
