% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeSurface.R
\docType{data}
\name{WholeSurface}
\alias{WholeSurface}
\title{Surface mesh of whole auricular surface.}
\format{
An object of class \code{mesh3d} of length 4.
}
\usage{
data(WholeSurface)
}
\description{
Surface mesh of whole auricular surface.
}
\examples{
PCQDA_output <- PCQDA_result(WholeSurface,Apex)
PCR_output <- PCR_result(WholeSurface,Apex)
PCLR_output <- PCLR_result(WholeSurface,Apex)
}
\keyword{datasets}
