% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PCR.R
\docType{data}
\name{data_PCR}
\alias{data_PCR}
\title{Database to create DNE_PCR method}
\format{
An object of class \code{data.frame} with 599 rows and 8 columns.
}
\usage{
data_PCR
}
\description{
data_PCR with 4 variables: Age, Median value of DNE on Apex, Area having Macroporosity, IQR of DNE on Whole surface
}
\details{
The database was obtained by calculating the DNE on the auricular surface from European collections.
The specimens are belongs to  the Louis Lopes collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan.
}
\keyword{datasets}
