% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSDNE-package.R
\docType{package}
\name{JSDNE-package}
\alias{JSDNE}
\alias{JSDNE-package}
\title{JSDNE: Estimating the age using auricular surface by DNE}
\description{
The package automatically estimates the age by calculating the DNE on the whole auricular surface and the apex of the auricular surface. It involves three estimation methods: DNE_PCQDA, DNE_PCR, and DNE_PCLR methods. The package is created with the data from the Louis Lopes collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan.
}
\author{
\strong{Maintainer}: Jisun Jang \email{jisun.jang.19@ucl.ac.uk}

}
\keyword{internal}
