% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSDNE-package.R
\docType{package}
\name{JSDNE-package}
\alias{JSDNE}
\alias{JSDNE-package}
\title{JSDNE: Estimating the Age using Auricular Surface by DNE}
\description{
The age is estimated by calculating the Dirichlet Normal Energy (DNE) on the whole auricular surface and the apex of the auricular surface. It involves three estimation methods: principal component discriminant analysis (PCQDA), and principal component logistic regression analysis (PCLR) methods, principal component regression analysis with Southeast Asian (A_PCR), and principal component regression analysis with multipopulation (M_PCR). The package is created with the data from the Louis Lopes Collection in Lisbon, the 21st Century Identified Human Remains Collection in Coimbra, and the CAL Milano Cemetery Skeletal Collection in Milan, and the skeletal collection at Khon Kaen University (KKU) Human Skeletal Research Centre (HSRC), housed in the Department of Anatomy in the Faculty of Medicine at KKU in Khon Kaen.
}
\author{
\strong{Maintainer}: Jisun Jang \email{jisun.jang.19@ucl.ac.uk}

}
\keyword{internal}
