% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawData.R
\docType{data}
\name{RawData}
\alias{RawData}
\title{RawData}
\format{
An object of class \code{data.frame} with 890 rows and 15 columns.
}
\usage{
RawData
}
\description{
RawData includes the raw information of DNE variables and biological
information of auricular surfaces (age and sex) of 890 auricular surface.
It consists of Age, Cluster1, Cluster2, Collection, Sex, MeanDNE.Apex,
MedianDNE.Apex, IQRDNE.Apex, TotalDNE.TotalPolygonFaces, MedianDNE.Whole,
IQRDNE.Whole, MeanDNE.Convex, MeanDNE.Concave, Proportion.DNEunder0.0001 and
Proportion.DNEover0.6. Cluster1 and Cluster2 is the clustering information
for  PCLR and PCQDA, respectively.
}
\keyword{datasets}
